fix(ci): de-duplicate explicitFileList helper; use [[ in changed-files.sh
SonarCloud quality gate flagged 13.3% duplication on new code — the
explicitFileList() function (CHANGED_FILES env + argv parsing) was copied
verbatim into all 3 check-*.mjs scripts (types-location 66.7%, components
58.1%). Extract it to scripts/ci/lib/explicit-file-list.mjs and import it;
each script keeps its own extension/scope/ignore filtering. Single source now.

Also switch the 4 `[ ... ]` tests in changed-files.sh to `[[ ... ]]`
(SonarCloud shell rule). shellcheck clean; all scripts verified in both
full-scan and CHANGED_FILES scoped modes.
