+++ scripts/ci/check-components-per-file.mjs
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 import { readFileSync } from "node:fs";
 import { execFileSync } from "node:child_process";
+import { explicitFileList } from "./lib/explicit-file-list.mjs";
 
 const ALLOWLIST_PREFIXES = [
   "apps/web/src/components/ui/",
@@ -49,18 +50,6 @@ const SCANNED_ROOTS = [
 
 function inScope(path) {
   return path.endsWith(".tsx") && SCANNED_ROOTS.some((r) => path.startsWith(r));
-}
-
-// Explicit file list (CI diff-scoping): newline-separated CHANGED_FILES env
-// var OR non-flag argv entries. When provided, only those in-scope files are
-// checked. Otherwise fall back to a full `git ls-files` repo scan.
-function explicitFileList() {
-  const fromEnv = (process.env.CHANGED_FILES ?? "")
-    .split("\n")
-    .map((s) => s.trim())
-    .filter(Boolean);
-  const fromArgv = process.argv.slice(2).filter((a) => !a.startsWith("-"));
-  return [...fromEnv, ...fromArgv];
 }
 
 function getFiles() {
