+++ scripts/ci/check-file-sizes.mjs
@@ -13,6 +13,7 @@
  */
 import { readFileSync } from "node:fs";
 import { execFileSync } from "node:child_process";
+import { explicitFileList } from "./lib/explicit-file-list.mjs";
 
 const DEFAULT_LIMIT = 400;
 const RELAXED_LIMIT = 700;
@@ -74,19 +75,6 @@ const exemptFromHardCap = (p) =>
   NO_HARD_CAP_PATTERNS.some((rx) => rx.test(p));
 
 const SCANNED_EXTENSIONS = [".ts", ".tsx", ".js", ".jsx", ".mjs", ".cjs"];
-
-// Explicit file list (CI diff-scoping): newline-separated CHANGED_FILES env
-// var OR non-flag argv entries. When provided, only those files are checked —
-// still subject to this script's own extension + ignore rules. Otherwise fall
-// back to a full `git ls-files` repo scan.
-function explicitFileList() {
-  const fromEnv = (process.env.CHANGED_FILES ?? "")
-    .split("\n")
-    .map((s) => s.trim())
-    .filter(Boolean);
-  const fromArgv = process.argv.slice(2).filter((a) => !a.startsWith("-"));
-  return [...fromEnv, ...fromArgv];
-}
 
 function getFiles() {
   const explicit = explicitFileList();
